+++ lib/form_shield.dart
@@ -1,6 +1,7 @@
 export 'src/validator.dart';
 export 'src/validation_rule.dart';
 export 'src/validation_result.dart';
+export 'src/async_validation_state.dart';
 export 'src/rules/email_rule.dart';
 export 'src/rules/password_rule.dart';
 export 'src/rules/length_rule.dart';